refactor(runtime): extract peer coordination tranche 6 (#3145)
