+++ docs/plans/2026-02-08-production-service-roadmap.md
@@ -66,6 +66,9 @@ There is **zero async code** in the entire codebase. No tokio, no `async fn`, no
 - Post-roadmap hardening wave 4 runtime decomposition tranche 5 delivered:
   - Extracted recovery/rejoin guard orchestration from `crates/kamn-core/src/runtime.rs` into dedicated module `crates/kamn-core/src/runtime_recovery_guard.rs` with unchanged external behavior (Task #3129, Subtask #3130).
   - Module-ownership and docs-contract markers enforced by `crates/kamn-core/tests/runtime_module_extraction_contract.rs` and `crates/kamn-core/tests/runtime_network_docs.rs`.
+- Post-roadmap hardening wave 5 runtime decomposition tranche 6 delivered:
+  - Extracted peer lifecycle/queue/frame/proposal/runtime-wiring orchestration from `crates/kamn-core/src/runtime.rs` into dedicated module `crates/kamn-core/src/runtime_peer_coordination.rs` with unchanged external behavior (Task #3145, Subtask #3155).
+  - Module-ownership and docs-contract markers enforced by `crates/kamn-core/tests/runtime_module_extraction_contract.rs` and `crates/kamn-core/tests/runtime_network_docs.rs`.
 - Post-roadmap hardening wave 3 managed-signer startup live validation delivered:
   - Contract lane: `scripts/kolme/run_managed_signer_startup_live_validation_contract_lane.sh` and `scripts/kolme/test_run_managed_signer_startup_live_validation_contract_lane.sh` (Subtask #3067).
   - Contract report schema: `kamn.kolme.managed-signer-startup-live-validation-contract-report.v1`.
